transfer-contract: change mint to inter-contract
- Add `Mint` variant to `Call` enum, containing the stealth address, the
  value to mint, and a nonce.
- Change `add_balance` function to be callable publicly. This is done so
  balance can be added to a contract while setting up the initial state.
- Change `mint` to be an inter-contract-call - callable only by the
  transfer contract. The call pushes a note with the desired value and
  address to the state at the current height.
- Remove Dusk public spend key.
